Restructuring the es_reader so that the core logic is reusable in other readers
